Split BgpConfigManagerTest and BgpConfigTest into separate files.
